+++ src/_modules/files/files.model.ts
@@ -10,8 +10,11 @@ export interface IFile {
 export const FileBodyDTO: TSchema = t.Object({
     file: t.File()
 })
+export const FilesBodyDTO: TSchema = t.Object({
+    files: t.Array(t.File())
+})
 export const ImageBodyDTO: TSchema = t.Object({
-    file: t.File({ type: 'image/*' })
+    file: t.File({ type: "image" })
 })
 export const ImagesBodyDTO: TSchema = t.Object({
     files: t.Array(t.File({ type: 'image/*' }))
@@ -20,11 +23,28 @@ export interface IImageUpload {
     name: string;
     size: number;
     type: string;
-    'X-Amz-Meta-UserProfileId'?: string;
+    'X-Amz-Meta-UserId'?: string;
     'X-Amz-Meta-CreatedDate'?: string;
+    'X-Amz-Meta-Mode'?: string;
+    'X-Amz-Meta-Purpose'?: string;
+}
+export interface IFileUpload {
+    name: string;
+    size: number;
+    type: string;
+    'X-Amz-Meta-UserId'?: string;
+    'X-Amz-Meta-CreatedDate'?: string;
+    'X-Amz-Meta-Purpose'?: string;
 }
 
+/** No underscores */
 export enum BucketType {
-    PRODUCT="products",
-    USER="users"
+    PHOTOS="hello-photos",
+    FILES="hello-files",
+    USERS="hello-users"
+}
+export enum UploadMode {
+    MAIN="main",
+    THUMB="thumb",
+    BLUR="blur"
 }